220 detect default side effects (#230)
Close #220

We introduced a list of functions that are treated as default side
effects having impact on all objects. Those will be always returned with
`get_code`. Currently the list contains of `library`, `require` and
`data`. I wonder if this should be a parameter to `get_code` so that
this list can be extended and adjusted by future users?

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>